Refactor 'cmake' support and add 'cmake_clang'
